+++ CHANGELOG.md
@@ -33,6 +33,15 @@
 
 ## Version 26
 
+### v26.2.0
+
+- Ability to specify a custom name for a schema in the generated Documentation:
+  - Use the `.meta()` method on a schema with an `{ id }` as an argument;
+  - The `id` must be unique across all schemas used in your API;
+  - The feature proposed by [@arlyon](https://github.com/arlyon) 2 years ago, but the
+    implementation became possible only with Zod 4 and thanks to suggestions
+    from [@Upsilon-Iridani](https://github.com/Upsilon-Iridani).
+
 ### v26.1.0
 
 - Optimization to the memory consumption for your API: